Split footnotes module into submodules
